requirements.txt
openai
python-dotenv
pandas
tiktoken
scipy
beautifulsoup4
contentful
matplotlib
flake8
isort